테스트 데이터 및 앞선 리팩토링으로 인한 버그 반영. 프로퍼티 파일 보강

+++ model/src/main/java/sandbox9/thunderbolt/event/product/handler/ProductPriceEventHandler.java
@@ -24,9 +24,9 @@ public class ProductPriceEventHandler implements ProductEventHandler {
         }
         long beforePrice = sku.getSalePrice();
         if (EventCalculationType.PLUS.equals(event.getCalculationType())) {
-            sku.setSalePrice(beforePrice + (Long) event.getValue());
+            sku.setSalePrice(beforePrice + (Integer) event.getValue());
         } else if (EventCalculationType.MINUS.equals(event.getCalculationType())) {
-            sku.setSalePrice(beforePrice - (Long) event.getValue());
+            sku.setSalePrice(beforePrice - (Integer) event.getValue());
         }
         logger.info("Sku 가격 변경 내역[" + product.getProductId() + " - " + sku.getSkuId() + "]" + beforePrice +
                 " => " + sku.getSalePrice() + "(" + event.getCalculationType() + ", " + event.getValue() + ")");
